# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OCBP API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/clinical-assistants/get/clinical-assistants": {
            "get": {
                "tags": [
                    "AuxiliaresClinicos"
                ],
                "summary": "Get Clinical Assistants Info",
                "description": "Returns Get Clinical Assistants Info",
                "operationId": "ClinicalAssistants",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
        "/api/v1/clinical-assistants/get/patient-info-by-hab-code/{hab?}": {
            "get": {
                "tags": [
                    "AuxiliaresClinicos"
                ],
                "summary": "Get Patient Info by Hab Code",
                "description": "Returns Patient Info by Hab Code",
                "operationId": "getPatientInfoByHabCode",
                "parameters": [
                    {
                        "name": "hab?",
                        "in": "path",
                        "description": "Habitation Code - Opcional",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
        "/api/v1/clinical-assistants/patient/{patientdoc?}/type/{patientdoctype?}": {
            "get": {
                "tags": [
                    "AuxiliaresClinicos"
                ],
                "summary": "Get Patient Informations",
                "description": "Returns Patient Information",
                "operationId": "initialPatientInfoClinicalAssistant",
                "parameters": [
                    {
                        "name": "patientdoc?",
                        "in": "path",
                        "description": "Documento del Paciente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "patientdoctype?",
                        "in": "path",
                        "description": "Tipo de Documento del Paciente - CC, TI, RC, PE",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
        "/api/v1/clinical-assistants/get-services": {
            "get": {
                "tags": [
                    "AuxiliaresClinicos"
                ],
                "summary": "Get Services Informations",
                "description": "Returns Services Information",
                "operationId": "getServices",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
...
        "/api/v1/macna/patient/{patientdoc}/type/{patientdoctype}/information": {
            "get": {
                "tags": [
                    "Macna"
                ],
                "summary": "Get Patient Informations",
                "description": "Returns Patient Information",
                "operationId": "initialPatientInfo",
                "parameters": [
                    {
                        "name": "patientdoc",
                        "in": "path",
                        "description": "Documento del Paciente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "patientdoctype",
                        "in": "path",
                        "description": "Tipo de Documento del Paciente - CC, TI, RC, PE",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "type": "apikey",
                        "in": "header",
                        "name": "X-Authorization",
                        "X-Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "X-Authorization": {
                "type": "apiKey",
                "description": "EspartaApiKey",
                "name": "X-Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "X-Authorization": []
        }
    ]
}